image pipeline: move into own module
